# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: IQmeetEQ Workshop Companion API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sections/unlock:
    post:
      operationId: unlockSection
      tags: [sections]
      summary: Unlock a section with a code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnlockRequest"
      responses:
        "200":
          description: Section unlocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnlockResponse"
        "400":
          description: Invalid code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /workflow-maps:
    get:
      operationId: getWorkflowMaps
      tags: [workflows]
      summary: Get all workflow maps for current participant
      responses:
        "200":
          description: List of workflow maps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkflowMap"

    post:
      operationId: createWorkflowMap
      tags: [workflows]
      summary: Create a new workflow map
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkflowMapRequest"
      responses:
        "201":
          description: Workflow map created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowMap"
  /workflow-maps/{id}:
    get:
      operationId: getWorkflowMap
      tags: [workflows]
      summary: Get a workflow map
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Workflow map
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowMap"

    put:
      operationId: updateWorkflowMap
      tags: [workflows]
      summary: Update a workflow map
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkflowMapRequest"
      responses:
        "200":
          description: Workflow map updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowMap"

    delete:
      operationId: deleteWorkflowMap
      tags: [workflows]
      summary: Delete a workflow map
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    UnlockRequest:
      type: object
      properties:
        code:
          type: string
      required:
        - code
    UnlockResponse:
      type: object
      properties:
        sectionId:
          type: string
        sectionTitle:
          type: string
        alreadyUnlocked:
          type: boolean
      required:
        - sectionId
        - sectionTitle
        - alreadyUnlocked
...
    WorkflowMap:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        frequency:
          type: string
        currentSteps:
          type: array
          items:
            type: string
        redesignedSteps:
          type: array
          items:
            type: string
        verificationCheckpoints:
          type: string
        stopConditions:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - frequency
        - currentSteps
        - redesignedSteps
        - verificationCheckpoints
        - stopConditions
        - createdAt
        - updatedAt
    CreateWorkflowMapRequest:
      type: object
      properties:
        name:
          type: string
        frequency:
          type: string
        currentSteps:
          type: array
          items:
            type: string
        redesignedSteps:
          type: array
          items:
            type: string
        verificationCheckpoints:
          type: string
        stopConditions:
          type: string
      required:
        - name
    UpdateWorkflowMapRequest:
      type: object
      properties:
        name:
          type: string
        frequency:
          type: string
        currentSteps:
          type: array
          items:
            type: string
        redesignedSteps:
          type: array
          items:
            type: string
        verificationCheckpoints:
          type: string
        stopConditions:
          type: string
...